+++ src/commands/retrieve.ts
@@ -111,7 +111,7 @@ export interface ToolingType {
   members: string[];
 }
 
-export interface ToolingTypes {
+interface ToolingTypes {
   types: ToolingType[];
 }
 
@@ -617,6 +617,17 @@ function getAnyNameFromUri(uri: vscode.Uri): Promise<PXMLMember> {
     const isAura: boolean = ffNameParts[0] === 'aura' || ffNameParts[0] === 'lwc';
     const isDir: boolean = fs.lstatSync(uri.fsPath).isDirectory();
     const tType: IMetadataObject | undefined = getAnyTTMetadataFromPath(uri.fsPath);
+    const isResource: boolean = ffNameParts[0] === 'resource-bundles';
+    if (isResource) {
+      if (ffNameParts.length === 1) {
+        // refresh all bundles
+        return resolve({ name: 'StaticResource', members: ['*'] });
+      } else {
+        // refresh specific bundle
+        const members: string = ffNameParts[1].split('.resource').shift() || '*';
+        return resolve({ name: 'StaticResource', members: [members] });
+      }
+    }
     if (!tType) {
       return reject(
         "Either the file/metadata type doesn't exist in the current org or you're trying to save/retrieve outside of " +
@@ -628,24 +639,24 @@ function getAnyNameFromUri(uri: vscode.Uri): Promise<PXMLMember> {
       // we have foldered metadata
       ffNameParts.shift();
       folderedName = ffNameParts.join('/').split('.')[0];
-      resolve({ name: tType.xmlName, members: [folderedName] });
+      return resolve({ name: tType.xmlName, members: [folderedName] });
     } else if (isDir) {
       if (isAura) {
         if (baseDirectoryName === 'aura' || baseDirectoryName === 'lwc') {
           baseDirectoryName = '*';
         }
-        resolve({ name: tType.xmlName, members: [baseDirectoryName] });
+        return resolve({ name: tType.xmlName, members: [baseDirectoryName] });
       } else if (tType.inFolder && ffNameParts.length > 1) {
-        getFolderContents(tType.xmlName, ffNameParts[1]).then(contents => {
-          resolve({ name: tType.xmlName, members: contents });
+        return getFolderContents(tType.xmlName, ffNameParts[1]).then(contents => {
+          return resolve({ name: tType.xmlName, members: contents });
         });
       } else {
-        getMembers([tType.xmlName], true).then(members => {
-          resolve(members[0]);
+        return getMembers([tType.xmlName], true).then(members => {
+          return resolve(members[0]);
         });
       }
     } else {
-      resolve({ name: tType.xmlName, members: [baseDirectoryName] });
+      return resolve({ name: tType.xmlName, members: [baseDirectoryName] });
     }
   });
 }