+++ src/forceCode.d.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import { Connection } from 'jsforce';
-import { ForceService } from './services';
+import { ForceService, SaveResult } from './services';
 
 declare module 'vscode' {
   export namespace window {
@@ -155,10 +155,11 @@ export interface IForceService {
   projectRoot: string;
   workspaceRoot: string;
   storageRoot: string;
-  describe?: IMetadataDescribe;
+  describe: IMetadataDescribe;
   conn: Connection;
   fcDiagnosticCollection: vscode.DiagnosticCollection;
   uuid: string;
+  lastSaveResult: SaveResult | undefined;
   connect(): Promise<IForceService>;
   checkForFileChanges(): any;
 }